+++ CHANGES.rst
@@ -9,6 +9,7 @@ development
 - Improve command line option juggling and documentation
 - Properly retry failing tasks with Celery
 - Honor "--use-application-id" also when printing results to STDOUT in bulk mode
+- Fix Celery task reject behavior
 
 2017-10-22 0.4.0
 ================